Rewrite dashboard/App.jsx to supress Redux error in hot reload
